Helpers `redirect()`, `redirectTo` and `reload` was moved to `Proxy\Response`
