+++ tests/src/Proxy/ResponseTest.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * @copyright Bluz PHP Team
+ * @link https://github.com/bluzphp/skeleton
+ */
+
+/**
+ * @namespace
+ */
+namespace Bluz\Tests\Proxy;
+
+use Bluz\Proxy\Response;
+use Bluz\Proxy\Router;
+use Bluz\Tests\TestCase;
+
+/**
+ * ResponseTest
+ *
+ * @package  Bluz\Tests\Proxy
+ * @author   Anton Shevchuk
+ */
+class ResponseTest extends TestCase
+{
+    /**
+     * Test Helper Redirect
+     *
+     * @expectedException \Bluz\Application\Exception\RedirectException
+     */
+    public function testHelperRedirect()
+    {
+        Response::redirect('/');
+    }
+
+    /**
+     * Test Helper RedirectTo
+     *
+     * @expectedException \Bluz\Application\Exception\RedirectException
+     */
+    public function testHelperRedirectTo()
+    {
+        Response::redirectTo(Router::getDefaultModule(), Router::getDefaultController());
+    }
+
+    /**
+     * Test Helper Reload
+     *
+     * @expectedException \Bluz\Application\Exception\ReloadException
+     */
+    public function testHelperReload()
+    {
+        Response::reload();
+    }
+}
